pull out styled input to its own file
